Split json config files
