+++ config/json/log.go
@@ -1,0 +1,9 @@
+package json
+
+type LogConfig struct {
+	AccessLogValue string `json:"access"`
+}
+
+func (config *LogConfig) AccessLog() string {
+	return config.AccessLogValue
+}
